Move utils.kt content to CameraMode.kt
